% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cup_journals}
\alias{cup_journals}
\title{List available journals}
\usage{
cup_journals()
}
\value{
A character vector of the journal names.
}
\description{
List available journal in the CUP template.
}
\examples{
CUPrticles::cup_journals()
}
