% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{journals}
\alias{journals}
\title{List available journals}
\usage{
journals()
}
\value{
A character vector of the journal names.
}
\description{
List available journal names in this package.
}
\examples{
CUPrticles::journals()
}
