% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{cup_article}
\alias{cup_article}
\title{R Markdown output formats for Cambridge University Press journals.
Adapted from
\url{https://www.overleaf.com/latex/templates/template-for-submission-to-political-science-research-and-methods/qytgqmgznvnh}.}
\usage{
cup_article(
  journal = c("default", "aog", "jog", "bjps", "psrm", "jpup", "jjps", "pasa"),
  manuscript = c("article", "communication", "letter", "note", "review", "suppinfo",
    "research-paper"),
  keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"),
  pandoc_args = NULL,
  ...
)
}
\arguments{
\item{journal}{The argument \code{journal} accepts the acronym of any of these
Cambridge University Press journals:
\itemize{
\item \code{default}: Default journal style
\item \code{aog}: Annals of Glaciology
\item \code{jog}: Journal of Glaciology
\item \code{bjps}: British Journal of Political Science
\item \code{psrm}: Political Science and Research Methods
\item \code{jpup}: Journal of Public Policy
\item \code{jjps}: Japanese Journal of Political Science
\item \code{pasa}: Publications of the Astronomical Society of Australia.
}}

\item{manuscript}{The argument \code{manuscript} accepts these types of journal contributions:
\itemize{
\item article
\item communication
\item letter
\item note
\item review
\item suppinfo
\item research-paper
}}

\item{..., number_sections, keep_tex, latex_engine, highlight, fig_caption, md_extensions, template, pandoc_args}{Arguments passed to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}
}
\description{
R Markdown output formats for Cambridge University Press journals.
Adapted from
\url{https://www.overleaf.com/latex/templates/template-for-submission-to-political-science-research-and-methods/qytgqmgznvnh}.
}
