% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{journals}
\alias{journals}
\title{List available templates}
\usage{
journals()
}
\value{
A character vector of the template names.
}
\description{
List available template names in this package.
}
\examples{
CUPrticles::journals()
}
